Merging in the latest JDK7 link.
According to archive.org, Oracle always seem to use
http://www.oracle.com/technetwork/java/javase/downloads/jdk7-downloads-1880260.html
for the latest JDK7 download, and this page has the same "downloads"
hash that the archive download page has. So we just need to merge in two
"downloads" hashes from two pages to compute the full JDK7 list,
including the very latest.

As JDK8 gets released, this will likely have to change.