Fix .NET SDK runtime info retrieval (#95)
Microsoft changed the tree structure for dotnet/runtime, breaking the
download. The URL has been updated to the new/current location.

In addition, some error handling was added, so that failure for one
of the three output files does not prevent creation of the others. An
aggregate exception will be thrown at the end.