Fixed duplicate id for sonar-scanner-msbuild versions
The PR supplied in #96 uses the same ID for two different versions. I'm actually curious how the initial PR ever worked, but it doesn't do so now.

I added a unique ID for the net5.0 version, which should fix this.

I also changed the display name to better reflect the official name of .NET 5.0 which MS promotes for it.