Check out `main` branch for jenkins-scripts
When checking out the jenkins-scripts repository for the Scriptler
catalog, check out from the `main` branch instead of the `master` branch
going forward. This change also ensures that all changes should be able
to be checked out, regardless of whether a rebase from the upstream repo
is possible or not.